Fixed rho_final value errors
Adapted for loop from traffic.py into advection() and now the output rho_final values match the output rplot values!
Had to remove old fix for the broadcasting error as now a new dimension wouldn't be necessary and the values are updated with a for loop.
Now testing plot for the first time.