Fixed t_intervals bug
Through comparing the t_intervals[18][:] and tp[1422:1501] I found that the final row of t_intervals wasn't being populated by the for loop. It didn't take long to notice that this was due to a logical error in my selected looping range, and after changing to 19 it worked as intended.